packages/aetherflow-core/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aetherflow-core"
version = "0.1.1"
description = "Workflow operations engine (flow/job/step) with state, resume, cleanup and plugins."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="aetherflow contributors"}]
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
]

[project.optional-dependencies]
zip = ["pyzipper>=0.3.6"]
rest = ["httpx>=0.27", "anyio>=4.0", "tenacity>=8.2"]
sftp = ["paramiko>=3.4"]
smb  = ["pysmb>=1.2.9", "smbprotocol>=1.11"]
db   = ["sqlalchemy>=2.0", "oracledb>=2.0", "psycopg2-binary>=2.9", "pymysql>=1.1", "pyexasol>=0.25"]
mail = []
duckdb = ["duckdb>=1.0"]
parquet = ["pyarrow>=16.0"]
excel = ["openpyxl>=3.1"]
reports = ["duckdb>=1.0", "pyarrow>=16.0", "openpyxl>=3.1"]

all  = ["pyzipper>=0.3.6", "httpx>=0.27", "anyio>=4.0", "tenacity>=8.2", "paramiko>=3.4", "pysmb>=1.2.9", "smbprotocol>=1.11", "sqlalchemy>=2.0", "oracledb>=2.0", "psycopg2-binary>=2.9", "pymysql>=1.1", "pyexasol>=0.25", "duckdb>=1.0", "pyarrow>=16.0", "openpyxl>=3.1"]

dev = ["pytest>=8.0", "pytest-timeout>=2.2", "pytest-xdist>=3.6", "pytest-cov>=5.0", "respx>=0.21", "ruff>=0.6", "openpyxl>=3.1", "pyzipper>=0.3.6"]

[project.scripts]
aetherflow = "aetherflow.core.cli:main"

[project.entry-points."aetherflow.plugins"]
# Third-party plugins can register here:
# myplugin = "myplugin_aetherflow.plugin:register"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true
include = ["aetherflow.core*"]

[tool.pytest.ini_options]
markers = [
  "contract: validates that shipped docs/demo examples match the FlowSpec contract",
]
